drop wrapper
fold Program into ProgramContext

cleanup
